Separate web and tauri download file utils
