+++ application/ui/src/shared/util.ts
@@ -1,8 +1,10 @@
-// Copyright (C) 2025 Intel Corporation
+// Copyright (C) 2025-2026 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
 import { isEmpty, isString } from 'lodash-es';
 import prettyBytes from 'pretty-bytes';
+
+import { downloadFile as platformDownloadFile } from '../platform/download-file';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type GetElementType<T extends any[]> = T extends (infer U)[] ? U : never;
@@ -13,17 +15,8 @@ export const isNonEmptyArray = <T>(value: T): value is IsValidArrayType<T> => Ar
 
 export const isNonEmptyString = (value: unknown): value is string => isString(value) && value !== '';
 
-export const downloadFile = (url: string, name?: string) => {
-    const link = document.createElement('a');
-
-    link.href = url;
-    if (name !== undefined) {
-        link.download = name;
-    }
-    link.hidden = true;
-    link.click();
-
-    URL.revokeObjectURL(url);
+export const downloadFile = (url: string, name?: string): void => {
+    platformDownloadFile(url, name);
 };
 
 export const formatBytes = (bytes: number): string => prettyBytes(bytes);
